requirements.txt
#python-redis
redis
flask
werkzeug
pyodbc
azure-storage-queue==1.4.0

#azure-storage-queue